Made data serializer coversions: - DateTime/DateTimeOffset to unix miliseconds, - bool to number optional. Can be enabled when configuring serializer.
Reason of the change is that no matter if conversions are applied or not, queries can be used in a way that's correct according to type system, but incorrect according to Druid. While the number of possible invalid queries with conversions enabled (with data stored in Druid matching conversions) is reduced, conversions themselves can be surprising to users. Making enabling them explicit removes the surprise.